Allow the creation of SelfAttestedClaims
These are claims that do not require a 2-party system to be correctly
generated. Essentially they place the trust on the attester that
everything will be correct and validated.

They require, however, additional security concerns to ensure that a
user does not generate credentials on behalf of a third user. This means
that these credentials are vulnerable to identity theft: the attester
will generate and attest a valid claim without relying on the
information that the claimer indeed owns the associated address. This
check should be ensured by the system that makes use of
SelfAttestedClaims.